Data from the Biodiversity-ecosystem functioning experiment in the Koffler Scientific Reserve (KSR)

Species composition and ecosystem functions in each plot. The phylogeny of the fourteen species in the plots is also provided.

Usage

data(KSR)
data(KSR_MLtree)
data(KSR_EF)

Format

These are linked datasets. KSR provides species composition data in 88 plots of 14 plant species, while KS_MLtree describes the phylogenetic relationship between the species. KSR_EF are the ten ecosystem functions measured in each plot.

The meaning of each column in KSR_EF is explained below:

Plot:

Plot identity. This can be ignored. This column also applies to KSR even though it is absent in the compositional data.

Real.rich:

Plant species richness in the plot.

litter2012:

Amount of litter measured in 2012.

ave.biomass:

Average biomass across 2012-2014.

LAI:

Leaf area index, a simplified dimension of structural complexity.

mean.N.change:

delta 15N change averaged across surface and deep soil.

poll_total:

Total number of pollinators.

flwr_total:

Total number of flowers

Mass.loss.2month:

Decomposition after 2 months.

Damage_effect:

Damage reduction effect.

bugs:

Total number of arthropods.

bug.rich:

Species richness of arthropods.

Details

Note that the compositional data is binary and describe whether the species is planted in each plot initally. The phylogenetic tree is constructed using Maximum Likelihood. See Cadotte et al. (2013) for details

Cadotte, M. W. (2013). Experimental evidence that evolutionarily diverse assemblages result in higher productivity. Proceedings of the National Academy of Sciences, 110(22), 8996-9000.

Author

Toby P.N. Tsang